Fixed and created parkingPriceTabWidget
